# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Menambahkan produk baru (ENTRY)",
        "tags": [
          "Products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for product name or SKU",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination (id_product)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "is_active",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mendapatkan daftar produk beserta stok terkini (Semua Role)",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update detail produk (ENTRY)",
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete produk (ENTRY)",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}/archive": {
      "post": {
        "operationId": "ProductsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}/restore": {
      "post": {
        "operationId": "ProductsController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}/stock": {
      "post": {
        "operationId": "ProductsController_adjustStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustStockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Penyesuaian stok manual / Opname (ENTRY)",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}/stock-adjustments": {
      "post": {
        "operationId": "ProductsController_adjustStockCanonical",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustStockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}/stock-history": {
      "get": {
        "operationId": "ProductsController_stockHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Products"
        ]
      }
    },
...
    "/api/v1/sync": {
      "post": {
        "operationId": "SyncController_accept",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncBatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Durably queue an offline transaction batch; 200 does not mean settled",
        "tags": [
          "Sync"
        ]
      }
    },
    "/api/v1/sync/receipts": {
      "get": {
        "operationId": "SyncController_getReceipts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Sync"
        ]
      }
    },
    "/api/v1/sync/failures": {
      "get": {
        "operationId": "SyncController_getFailures",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List terminal sync failures and stock conflicts for Owner action",
        "tags": [
          "Sync"
        ]
      }
    },
    "/api/v1/sync/receipts/{id}/retry": {
      "post": {
        "operationId": "SyncController_retry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Sync"
        ]
      }
    },
...
    "/api/v1/owner/dashboard": {
      "get": {
        "operationId": "ReportingController_dashboard",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Owner"
        ]
      }
    },
    "/api/v1/audit-events": {
      "get": {
        "operationId": "ReportingController_auditEvents",
        "parameters": [
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Owner"
        ]
      }
    }
  },
  "info": {
    "title": "K-POS API",
    "description": "Canonical API for the K-POS multi-app offline-first platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nama produk",
            "example": "Mie Goreng"
          },
          "sku": {
            "type": "string",
            "description": "SKU produk",
            "example": "MG-001"
          },
          "price": {
            "type": "number",
            "description": "Harga produk",
            "example": 15000
          },
          "is_active": {
            "type": "boolean",
            "description": "Status aktif",
            "example": true
          },
          "image": {
            "type": "string",
            "format": "binary",
            "description": "Gambar produk (opsional, maks 5MB)"
          }
        },
        "required": [
          "name",
          "price"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nama produk",
            "example": "Mie Goreng"
          },
          "sku": {
            "type": "string",
            "description": "SKU produk",
            "example": "MG-001"
          },
          "price": {
            "type": "number",
            "description": "Harga produk",
            "example": 15000
          },
          "is_active": {
            "type": "boolean",
            "description": "Status aktif",
            "example": true
          },
          "image": {
            "type": "string",
            "format": "binary",
            "description": "Gambar produk (opsional, maks 5MB)"
          }
        }
      },
      "AdjustStockDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "description": "Quantity to adjust (positive for addition, negative for reduction)",
            "example": 50
          },
          "notes": {
            "type": "string",
            "description": "Notes for the stock adjustment",
            "example": "Barang masuk dari supplier"
          }
        },
        "required": [
          "quantity"
        ]
      },
...
      "SyncItemDto": {
        "type": "object",
        "properties": {
          "id_product": {
            "type": "string",
            "example": "clprdxxx...",
            "description": "CUID of the product"
          },
          "product_name": {
            "type": "string",
            "example": "Kopi Susu Aren"
          },
          "product_sku": {
            "type": "string",
            "example": "KSA-01"
          },
          "catalog_version": {
            "type": "number",
            "example": 7
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Quantity purchased"
          },
          "unit_price": {
            "type": "number",
            "example": 15000,
            "description": "Price per unit at the time of transaction"
          },
          "subtotal": {
            "type": "number",
            "example": 30000,
            "description": "Subtotal for this item (quantity * unit_price)"
          }
        },
        "required": [
          "id_product",
          "product_name",
          "product_sku",
          "catalog_version",
          "quantity",
          "unit_price",
          "subtotal"
        ]
      },
      "SyncPaymentDto": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "CASH",
              "STATIC_QRIS",
              "BANK_TRANSFER"
            ],
            "example": "CASH"
          },
          "amount": {
            "type": "number",
            "example": 30000,
            "description": "Total amount to be paid"
          },
          "cash_received": {
            "type": "number",
            "example": 50000,
            "description": "Amount of cash received from customer"
          },
          "change_amount": {
            "type": "number",
            "example": 20000,
            "description": "Change amount given to customer"
          },
          "qris_code": {
            "type": "string",
            "example": "QRIS-12345",
            "description": "QRIS reference code if applicable"
          },
          "transfer_ref": {
            "type": "string",
            "example": "TRX-BANK-001",
            "description": "Bank transfer reference if applicable"
          }
        },
        "required": [
          "method",
          "amount"
        ]
      },
      "SyncTransactionDto": {
        "type": "object",
        "properties": {
          "offline_uuid": {
            "type": "string",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479",
            "description": "Locally generated UUID v4 or v7 for idempotency"
          },
          "created_at_local": {
            "type": "string",
            "example": "2026-08-16T10:00:00.000Z",
            "description": "ISO-8601 timestamp of when the transaction occurred locally"
          },
          "subtotal": {
            "type": "number",
            "example": 30000,
            "description": "Subtotal of all items"
          },
          "total": {
            "type": "number",
            "example": 30000,
            "description": "Total after discounts/taxes (currently same as subtotal)"
          },
          "notes": {
            "type": "string",
            "example": "Customer in a rush",
            "description": "Optional transaction notes"
          },
          "items": {
            "description": "List of items purchased",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncItemDto"
            }
          },
          "payment": {
            "description": "Payment details for this transaction",
            "allOf": [
              {
                "$ref": "#/components/schemas/SyncPaymentDto"
              }
            ]
          }
        },
        "required": [
          "offline_uuid",
          "created_at_local",
          "subtotal",
          "total",
          "items",
          "payment"
        ]
      },
      "SyncBatchDto": {
        "type": "object",
        "properties": {
          "transactions": {
            "description": "Batch of offline transactions to synchronize (max 100)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncTransactionDto"
            }
          }
        },
        "required": [
          "transactions"
        ]
      }
    }
  }
}